handle selecting current config, option to launch soundmodem on exit

+++ switch-soundmodem-configs.ps1
@@ -12,22 +12,36 @@ $configs | ForEach-Object {
     $count++ 
 }
 
-$selectedConfig = Read-Host "Select a config"
+$selectedConfig = Read-Host "`nSelect a config number"
 # Write-Host $configs[$selectedConfig]
 if(Test-Path $configs[$selectedConfig]) {
-    if(Test-Path ".\soundmodem.ini") {
-        Write-Host "Deleting soundmodem.ini"
-        Remove-Item "soundmodem.ini"
+    if($configs[$selectedConfig].name -eq "soundmodem.ini") {
+        Write-Host "soundmodem.ini was selected. Nothing to do."
+    }   
+    else {
+        if(Test-Path ".\soundmodem.ini") {
+            Write-Host "Deleting soundmodem.ini"
+            Remove-Item "soundmodem.ini"
+        }
+        else { Write-Host -ForegroundColor Yellow "soundmodem.ini does not exist" }
+
+        Write-Host "Copying $($configs[$selectedConfig].name) to soundmodem.ini"
+        Copy-Item $configs[$selectedConfig] -Destination ".\soundmodem.ini"
     }
-    else { Write-Host -ForegroundColor Yellow "soundmodem.ini does not exist" }
-    
-    Write-Host "Copying $($configs[$selectedConfig].name) to soundmodem.ini"
-    Copy-Item $configs[$selectedConfig] -Destination ".\soundmodem.ini"
 }
 else { 
     Write-Host -ForegroundColor Red "$($selectedConfig) not found. Exiting." 
     exit 
 }
 
-Read-Host -prompt "Press enter to exit"
+$launchSoundmodem = Read-Host "`nLaunch SoundModem? (y/n)" 
+if($launchSoundmodem -eq "y" -or $launchSoundmodem -eq "yes") {
+    if(Test-Path ".\soundmodem.exe") {
+        Start-Process ".\soundmodem.exe"
+    }
+    else {
+        Write-Host -ForegroundColor Red "soundmodem.exe not found. Exiting."
+    }
+}
+
 exit
